requirements.txt
flask
gunicorn
python-dotenv
os
google.generativeai